Upgrade appserver

+++ appservers/php/start.sh
@@ -14,7 +14,7 @@ if [ -f "/src/config/php/php.ini" ]; then
 fi
 
 # Use our custom apc.ini
-if [ -f "/src/config/php/apc.ini" ]; then
+if [ -f "/src/config/php/apc.ini" && -f "${HOME}/.phpbrew/php/php-${PHP_VERSION}/var/db/apc.ini"]; then
   ~/.phpbrew/php/php-5.3.29/var/db
   rm ${HOME}/.phpbrew/php/php-${PHP_VERSION}/var/db/apc.ini
   cp /src/config/php/apc.ini ${HOME}/.phpbrew/php/php-${PHP_VERSION}/var/db/apc.ini
